patch

+++ packages/react-native/ReactCommon/react/renderer/core/RawPropsParser.cpp
@@ -69,7 +69,7 @@ const RawValue* RawPropsParser::at(
   do {
     rawProps.keyIndexCursor_++;
 
-    if (UNLIKELY(rawProps.keyIndexCursor_ >= keys_.size())) {
+    if (UNLIKELY(static_cast<size_t>(rawProps.keyIndexCursor_) >= keys_.size())) {
 #ifdef REACT_NATIVE_DEBUG
       if (resetLoop) {
         LOG(ERROR)
